test(amplify-e2e-tests): fixes auth test
